+++ packages/mongodb-access-adapter/datagrip-access-adapter/src/main/kotlin/com/mongodb/jbplugin/accessadapter/datagrip/adapter/DataGripMongoDbDriver.kt
@@ -127,7 +127,7 @@ class DataGripMongoDbDriver(
                 val listOfResults = mutableListOf<T>()
                 val queryResult = runCatching { statement.executeQuery() }
                 if (queryResult.isFailure) {
-                    logger.error(
+                    logger.warn(
                         "Can not query MongoDB: $queryString",
                         queryResult.exceptionOrNull()
                     )