+++ packages/mongodb-mql-engines/src/main/kotlin/com/mongodb/jbplugin/indexing/IndexAnalyzer.kt
@@ -1,21 +1,14 @@
 package com.mongodb.jbplugin.indexing
 
 import com.mongodb.jbplugin.mql.BsonAny
-import com.mongodb.jbplugin.mql.BsonType
-import com.mongodb.jbplugin.mql.CollectionSchema
 import com.mongodb.jbplugin.mql.Node
 import com.mongodb.jbplugin.mql.SiblingQueriesFinder
 import com.mongodb.jbplugin.mql.components.HasCollectionReference
-import com.mongodb.jbplugin.mql.components.HasFieldReference
+import com.mongodb.jbplugin.mql.components.HasFilter
 import com.mongodb.jbplugin.mql.components.Name
+import com.mongodb.jbplugin.mql.components.Named
 import com.mongodb.jbplugin.mql.components.QueryRole
 import com.mongodb.jbplugin.mql.parser.*
-import com.mongodb.jbplugin.mql.parser.components.aggregationStages
-import com.mongodb.jbplugin.mql.parser.components.allNodesWithSchemaFieldReferences
-import com.mongodb.jbplugin.mql.parser.components.extractOperation
-import com.mongodb.jbplugin.mql.parser.components.extractValueReferencesRelevantForIndexing
-import com.mongodb.jbplugin.mql.parser.components.fieldReference
-import com.mongodb.jbplugin.mql.parser.components.hasName
 
 /**
  * The index analyzer is responsible for processing a query and return an index
@@ -46,7 +39,8 @@ object IndexAnalyzer {
         options: CollectionIndexConsolidationOptions
     ): SuggestedIndex<S> {
         val baseIndex = guessIndexForQuery(query)
-        val otherIndexes = siblingQueriesFinder.allSiblingsOf(query).map { guessIndexForQuery(it) }
+        val siblingQueries = siblingQueriesFinder.allSiblingsOf(query)
+        val otherIndexes = siblingQueries.map { guessIndexForQuery(it) }
 
         return CollectionIndexConsolidation.apply(baseIndex, otherIndexes, options)
     }
@@ -57,13 +51,14 @@ object IndexAnalyzer {
 
         val schema = (collectionRef.reference as? HasCollectionReference.Known<S>)?.schema
 
-        val allFieldUsages = query.allFieldReferences(schema).groupBy(QueryFieldUsage<S>::fieldName)
+        val allFieldUsages = IndexQueryFieldUsage.allFieldReferences(query, schema)
+            .groupBy(IndexQueryFieldUsage<S>::fieldName)
             .mapValues { (_, usages) -> usages.sortedBy { it.role.ordinal } }
 
         val contextInferredFieldUsages = allFieldUsages.mapValues { (_, usages) ->
             val leastSpecificUsage = usages
                 .filter { it.role == QueryRole.EQUALITY || it.role == QueryRole.RANGE }
-                .reduceOrNull(QueryFieldUsage<S>::leastSpecificUsageOf)
+                .reduceOrNull(IndexQueryFieldUsage<S>::leastSpecificUsageOf)
 
             // There might be more than one sort direction specified in the code
             // so we take only the last defined direction.
@@ -78,8 +73,26 @@ object IndexAnalyzer {
             )
         }
 
+        val queryRefForFilterExpression = contextInferredFieldUsages.values.mapNotNull {
+            it.reference
+        }
+
+        val partialFilterExpression = if (queryRefForFilterExpression.isEmpty()) {
+            null
+        } else if (queryRefForFilterExpression.size == 1) {
+            queryRefForFilterExpression.first()
+        } else {
+            Node(
+                Unit,
+                listOf(
+                    Named(Name.AND),
+                    HasFilter(queryRefForFilterExpression)
+                )
+            )
+        }
+
         val indexFields = contextInferredFieldUsages.values
-            .sortedWith(QueryFieldUsage.byRoleSelectivityAndCardinality())
+            .sortedWith(IndexQueryFieldUsage.byRoleSelectivityAndCardinality())
             .map {
                 SuggestedIndex.MongoDbIndexField(
                     source = it.source,
@@ -90,132 +103,20 @@ object IndexAnalyzer {
             }
             .toList()
 
-        return SuggestedIndex.MongoDbIndex(collectionRef, indexFields, listOf(query), null)
-    }
-
-    private suspend fun <S> Node<S>.allFieldReferences(
-        collectionSchema: CollectionSchema?
-    ): List<QueryFieldUsage<S>> {
-        val nodeQueryUsage = requireNonNull<Node<S>, NoConditionFulfilled>(NoConditionFulfilled)
-            .zip(extractOperation<S>())
-            .zip(extractValueReferencesRelevantForIndexing())
-            .zip(fieldReference<HasFieldReference.FromSchema<S>, S>())
-            .map { context ->
-                val named = context.first.first.second
-                val valueRef = context.first.second
-                val fieldRef = context.second
-                val role = named.queryRole(valueRef.type)
-                val fieldType = collectionSchema?.typeOf(fieldRef.fieldName) ?: BsonAny
-                val selectivity = if (role == QueryRole.SORT) {
-                    // In case of sort the selectivity is always null because there is no value to
-                    // calculate the selectivity from.
-                    null
-                } else {
-                    collectionSchema?.dataDistribution?.getSelectivityForPath(
-                        fieldRef.fieldName,
-                        valueRef.value,
-                    )
-                }
-
-                QueryFieldUsage(
-                    source = fieldRef.source,
-                    fieldName = fieldRef.fieldName,
-                    fieldType = fieldType,
-                    value = valueRef.value,
-                    valueType = valueRef.type,
-                    role = role,
-                    selectivity = selectivity,
-                    sortDirection = if (valueRef.value == -1) {
-                        SortDirection.Descending
-                    } else {
-                        SortDirection.Ascending
-                    }
-                )
-            }.anyError()
-
-        val extractAllFieldReferencesWithValues = allNodesWithSchemaFieldReferences<S>()
-            .mapMany(nodeQueryUsage)
-
-        val extractFromFirstMatchStage = aggregationStages<S>()
-            .nth(0)
-            .matches(hasName(Name.MATCH))
-            .flatMap(extractAllFieldReferencesWithValues)
-            .anyError()
-
-        val extractFromFiltersWhenNoAggregation = requireNonNull<Node<S>, Any>(Unit)
-            .matches(aggregationStages<S>().filter { it.isEmpty() }.matches())
-            .flatMap(extractAllFieldReferencesWithValues)
-            .anyError()
-
-        val findIndexableFieldReferences = first(
-            extractFromFirstMatchStage,
-            extractFromFiltersWhenNoAggregation,
+        return SuggestedIndex.MongoDbIndex(
+            collectionRef,
+            indexFields,
+            listOf(query),
+            partialFilterExpression as Node<S>?
         )
-
-        return findIndexableFieldReferences
-            .parse(this)
-            .orElse { emptyList() }
     }
 
-    private fun <S> reasonOfSuggestion(it: QueryFieldUsage<S>): IndexSuggestionFieldReason {
+    private fun <S> reasonOfSuggestion(it: IndexQueryFieldUsage<S>): IndexSuggestionFieldReason {
         return when (it.role) {
             QueryRole.EQUALITY -> IndexSuggestionFieldReason.RoleEquality
             QueryRole.SORT -> IndexSuggestionFieldReason.RoleSort
             QueryRole.RANGE -> IndexSuggestionFieldReason.RoleRange
             else -> IndexSuggestionFieldReason.RoleEquality
-        }
-    }
-
-    private data class QueryFieldUsage<S>(
-        val source: S,
-        val fieldName: String,
-        val fieldType: BsonType,
-        val value: Any?,
-        val valueType: BsonType,
-        val role: QueryRole,
-        val selectivity: Double?,
-        val sortDirection: SortDirection,
-    ) {
-        companion object {
-            fun <S> byRoleSelectivityAndCardinality() = Comparator<QueryFieldUsage<S>> { a, b ->
-                val roleComparison = a.role.ordinal.compareTo(b.role.ordinal)
-                if (roleComparison != 0) return@Comparator roleComparison
-
-                if (a.selectivity != null && b.selectivity != null) {
-                    val selectivityComparison = b.selectivity.compareTo(a.selectivity)
-                    if (selectivityComparison != 0) return@Comparator selectivityComparison
-                }
-
-                return@Comparator when {
-                    // In case of query role == Sort we compare by field types because value types
-                    // are ir-relevant as they are always inferred as 1 or -1.
-                    a.role == QueryRole.SORT && b.role == QueryRole.SORT ->
-                        a.fieldType.cardinality.compareTo(b.fieldType.cardinality)
-                    else ->
-                        a.valueType.cardinality.compareTo(b.valueType.cardinality)
-                }
-            }
-        }
-
-        /**
-         * For index suggestion, we want to promote a value that is least specific.
-         * This applies for cases when we have different values provided for a field (generally
-         * in an OR condition) in a query. If multiple values tie on specificity then we promote
-         * the one with high cardinality so we can ultimately force the field towards the end of
-         * the index definition.
-         */
-        fun leastSpecificUsageOf(other: QueryFieldUsage<S>): QueryFieldUsage<S> {
-            return if (other.value == null && this.value == null) {
-                if (this.valueType.cardinality > other.valueType.cardinality) {
-                    this
-                } else {
-                    other
-                }
-            } else if (other.value == null) {
-                other
-            } else {
-                this
-            }
         }
     }